Split library API rows and registry
