+++ crates/nose-semantics/src/library_api.rs
@@ -8,7 +8,14 @@ use super::*;
 use crate::evidence::span_contains;
 
 mod contracts;
+mod registry;
+mod rows;
 pub use contracts::*;
+use registry::{
+    library_api_callee_contract_for_hash, library_api_contract_id_from_hash,
+    library_api_contract_result_domain_for_arity, library_api_record_admitted_for_current_shape,
+};
+pub use rows::*;
 
 pub fn library_api_contract_evidence_for_call(
     il: &Il,
@@ -1869,518 +1876,6 @@ fn library_api_dependency_id_for_call_contract(
     found
 }
 
-fn library_api_contract_result_domain_for_arity(
-    id: LibraryApiContractId,
-    callee: LibraryApiCalleeContract,
-    arity: u16,
-) -> Option<DomainEvidence> {
-    match id {
-        LibraryApiContractId::PythonBuiltinCollectionFactory
-        | LibraryApiContractId::PythonImportedCollectionFactory
-        | LibraryApiContractId::RustStdCollectionFactory
-        | LibraryApiContractId::RustVecMacroFactory
-        | LibraryApiContractId::RustVecNewFactory
-        | LibraryApiContractId::JavaCollectionFactory(_)
-        | LibraryApiContractId::JavaCollectionConstructor(_)
-        | LibraryApiContractId::RubySetFactory
-        | LibraryApiContractId::JsLikeSetConstructor => {
-            library_collection_factory_result_domain_for_arity(
-                LibraryCollectionFactoryContract {
-                    id,
-                    callee,
-                    result: LibraryCollectionFactoryResult::SequenceArgument,
-                },
-                arity as usize,
-            )
-        }
-        LibraryApiContractId::RustStdMapFactory
-        | LibraryApiContractId::JavaMapFactory(_)
-        | LibraryApiContractId::JsLikeMapConstructor => Some(library_map_factory_result_domain(
-            LibraryMapFactoryContract {
-                id,
-                callee,
-                result: LibraryMapFactoryResult::EntrySequence {
-                    entry_seq_tag: SEQ_VALUE_COLLECTION,
-                },
-            },
-        )),
-        LibraryApiContractId::MapKeyViewWrapper => Some(
-            library_map_key_view_wrapper_result_domain(LibraryMapKeyViewWrapperContract {
-                id,
-                callee,
-                result: MapKeyViewWrapperContract {
-                    receiver: "Array",
-                    method: "from",
-                    qualified_path: "Array.from",
-                },
-            }),
-        ),
-        LibraryApiContractId::RustOptionSomeConstructor => Some(DomainEvidence::Option),
-        LibraryApiContractId::ScalarIntegerMethod(_) => Some(DomainEvidence::Integer),
-        LibraryApiContractId::MethodCall(MethodSemanticContract::HoF(_)) => {
-            Some(DomainEvidence::Collection)
-        }
-        _ => None,
-    }
-}
-
-fn library_api_contract_id_from_hash(hash: u64) -> Option<LibraryApiContractId> {
-    library_api_contract_ids()
-        .into_iter()
-        .find(|id| library_api_contract_id_hash(*id) == hash)
-}
-
-fn library_api_contract_ids() -> Vec<LibraryApiContractId> {
-    let mut ids = vec![
-        LibraryApiContractId::PropertyBuiltin(Builtin::Len),
-        LibraryApiContractId::PythonBuiltinCollectionFactory,
-        LibraryApiContractId::PythonImportedCollectionFactory,
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Len),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Append),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Print),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Range),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Sum),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Min),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Max),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Abs),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Zip),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Enumerate),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::Any),
-        LibraryApiContractId::FreeFunctionBuiltin(Builtin::All),
-        LibraryApiContractId::RustOptionSomeConstructor,
-        LibraryApiContractId::RustOptionNoneSentinel,
-        LibraryApiContractId::RustOptionAndThen,
-        LibraryApiContractId::RustStdCollectionFactory,
-        LibraryApiContractId::RustStdMapFactory,
-        LibraryApiContractId::RustVecMacroFactory,
-        LibraryApiContractId::RustVecNewFactory,
-        LibraryApiContractId::JavaMapEntryFactory,
-        LibraryApiContractId::RubySetFactory,
-        LibraryApiContractId::JsLikeSetConstructor,
-        LibraryApiContractId::JsLikeMapConstructor,
-        LibraryApiContractId::MapKeyViewWrapper,
-        LibraryApiContractId::MapGet,
-        LibraryApiContractId::JsArrayIsArray,
-        LibraryApiContractId::JsBooleanCoercion,
-        LibraryApiContractId::RegexTest,
-        LibraryApiContractId::JsLikeStaticIndexMembership(StaticIndexMembershipKind::IndexOf),
-        LibraryApiContractId::JsLikeStaticIndexMembership(StaticIndexMembershipKind::FindIndex),
-        LibraryApiContractId::PromiseThen,
-        LibraryApiContractId::IteratorIdentityAdapter,
-        LibraryApiContractId::StaticCollectionAdapter,
-    ];
-    ids.extend(
-        [
-            ScalarIntegerMethod::Abs,
-            ScalarIntegerMethod::Min,
-            ScalarIntegerMethod::Max,
-            ScalarIntegerMethod::Clamp,
-        ]
-        .into_iter()
-        .map(LibraryApiContractId::ScalarIntegerMethod),
-    );
-    ids.extend(
-        [
-            JavaCollectionFactoryKind::ListOf,
-            JavaCollectionFactoryKind::SetOf,
-            JavaCollectionFactoryKind::ArraysAsList,
-        ]
-        .into_iter()
-        .map(LibraryApiContractId::JavaCollectionFactory),
-    );
-    ids.push(LibraryApiContractId::JavaCollectionConstructor(
-        JavaCollectionConstructorKind::EmptyList,
-    ));
-    ids.extend(
-        [JavaMapFactoryKind::Of, JavaMapFactoryKind::OfEntries]
-            .into_iter()
-            .map(LibraryApiContractId::JavaMapFactory),
-    );
-    ids.extend(
-        [MapKeyViewKind::Collection, MapKeyViewKind::Iterator]
-            .into_iter()
-            .map(LibraryApiContractId::MapKeyView),
-    );
-    ids.extend(
-        [ImportedNamespaceFunctionSemantic::ProductReduction {
-            op: Op::Mul,
-            identity: 1,
-        }]
-        .into_iter()
-        .map(LibraryApiContractId::ImportedNamespaceFunction),
-    );
-    ids.extend(
-        [
-            MethodSemanticContract::Builtin(Builtin::Append),
-            MethodSemanticContract::Builtin(Builtin::Print),
-            MethodSemanticContract::Builtin(Builtin::Len),
-            MethodSemanticContract::Builtin(Builtin::IsEmpty),
-            MethodSemanticContract::Builtin(Builtin::IsNull),
-            MethodSemanticContract::Builtin(Builtin::IsNotNull),
-            MethodSemanticContract::Builtin(Builtin::StartsWith),
-            MethodSemanticContract::Builtin(Builtin::EndsWith),
-            MethodSemanticContract::Builtin(Builtin::Contains),
-            MethodSemanticContract::Builtin(Builtin::Join),
-            MethodSemanticContract::Builtin(Builtin::GetOrDefault),
-            MethodSemanticContract::Builtin(Builtin::ValueOrDefault),
-            MethodSemanticContract::Builtin(Builtin::Reduce),
-            MethodSemanticContract::Builtin(Builtin::Sum),
-            MethodSemanticContract::Builtin(Builtin::Abs),
-            MethodSemanticContract::Builtin(Builtin::Min),
-            MethodSemanticContract::Builtin(Builtin::Max),
-            MethodSemanticContract::Builtin(Builtin::Zip),
-            MethodSemanticContract::Builtin(Builtin::Any),
-            MethodSemanticContract::Builtin(Builtin::All),
-            MethodSemanticContract::HoF(HoFKind::Map),
-            MethodSemanticContract::HoF(HoFKind::Filter),
-            MethodSemanticContract::HoF(HoFKind::FlatMap),
-            MethodSemanticContract::HoF(HoFKind::FilterMap),
-        ]
-        .into_iter()
-        .map(LibraryApiContractId::MethodCall),
-    );
-    ids
-}
-
-fn library_api_record_admitted_for_current_shape(
-    il: &Il,
-    interner: &Interner,
-    call: NodeId,
-    record: &EvidenceRecord,
-) -> bool {
-    let EvidenceKind::LibraryApi(LibraryApiEvidenceKind::Contract {
-        contract_hash,
-        callee_hash,
-        arity,
-    }) = record.kind
-    else {
-        return false;
-    };
-    let Some(id) = library_api_contract_id_from_hash(contract_hash) else {
-        return false;
-    };
-    let Some(callee) = library_api_callee_contract_for_hash(il.meta.lang, id, callee_hash) else {
-        return false;
-    };
-    matches!(
-        library_api_contract_evidence_for_call(il, interner, call, id, callee, arity as usize),
-        LibraryApiEvidenceStatus::Admitted
-    )
-}
-
-fn library_api_callee_contract_for_hash(
-    lang: Lang,
-    id: LibraryApiContractId,
-    hash: u64,
-) -> Option<LibraryApiCalleeContract> {
-    library_api_callee_contracts_for_id(lang, id)
-        .into_iter()
-        .find(|callee| library_api_callee_contract_hash(*callee) == hash)
-}
-
-fn library_api_callee_contracts_for_id(
-    lang: Lang,
-    id: LibraryApiContractId,
-) -> Vec<LibraryApiCalleeContract> {
-    match id {
-        LibraryApiContractId::PropertyBuiltin(builtin) => ["length"]
-            .into_iter()
-            .filter_map(|property| library_property_builtin_contract(lang, property))
-            .filter(|contract| contract.id == LibraryApiContractId::PropertyBuiltin(builtin))
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::PythonBuiltinCollectionFactory
-        | LibraryApiContractId::RustStdCollectionFactory => {
-            library_free_name_collection_factory_contracts(lang)
-                .filter(|contract| contract.id == id)
-                .map(|contract| contract.callee)
-                .collect()
-        }
-        LibraryApiContractId::PythonImportedCollectionFactory => {
-            library_imported_collection_factory_contracts(lang)
-                .filter(|contract| contract.id == id)
-                .map(|contract| contract.callee)
-                .collect()
-        }
-        LibraryApiContractId::FreeFunctionBuiltin(builtin) => {
-            library_free_function_builtin_callee_contracts_for_id(lang, builtin)
-        }
-        LibraryApiContractId::RustOptionSomeConstructor => [
-            "Some",
-            "Option::Some",
-            "std::option::Option::Some",
-            "core::option::Option::Some",
-        ]
-        .into_iter()
-        .filter_map(|name| library_rust_option_some_constructor_contract(lang, name, 1))
-        .map(|contract| contract.callee)
-        .collect(),
-        LibraryApiContractId::RustOptionNoneSentinel => [
-            "None",
-            "Option::None",
-            "std::option::Option::None",
-            "core::option::Option::None",
-        ]
-        .into_iter()
-        .filter_map(|name| library_rust_option_none_sentinel_contract(lang, name))
-        .map(|contract| contract.callee)
-        .collect(),
-        LibraryApiContractId::RustOptionAndThen => {
-            library_rust_option_and_then_contract(lang, "and_then", 1)
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::ScalarIntegerMethod(method) => ["abs", "min", "max", "clamp"]
-            .into_iter()
-            .filter_map(|name| library_scalar_integer_method_contract(lang, name, 0))
-            .chain(
-                ["abs", "min", "max", "clamp"]
-                    .into_iter()
-                    .filter_map(|name| library_scalar_integer_method_contract(lang, name, 1)),
-            )
-            .chain(
-                ["abs", "min", "max", "clamp"]
-                    .into_iter()
-                    .filter_map(|name| library_scalar_integer_method_contract(lang, name, 2)),
-            )
-            .filter(|contract| contract.id == LibraryApiContractId::ScalarIntegerMethod(method))
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::RustStdMapFactory => library_free_name_map_factory_contracts(lang)
-            .filter(|contract| contract.id == id)
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::RustVecMacroFactory => {
-            library_rust_vec_macro_factory_contract(lang, "vec")
-                .filter(|contract| contract.id == id)
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::RustVecNewFactory => {
-            ["Vec::new", "std::vec::Vec::new", "alloc::vec::Vec::new"]
-                .into_iter()
-                .filter_map(|name| library_rust_vec_new_factory_contract(lang, name))
-                .filter(|contract| contract.id == id)
-                .map(|contract| contract.callee)
-                .collect()
-        }
-        LibraryApiContractId::JavaCollectionFactory(kind) => {
-            [("List", "of"), ("Set", "of"), ("Arrays", "asList")]
-                .into_iter()
-                .filter_map(|(receiver, method)| {
-                    library_java_collection_factory_contract(lang, receiver, method)
-                })
-                .filter(|contract| contract.id == LibraryApiContractId::JavaCollectionFactory(kind))
-                .map(|contract| contract.callee)
-                .collect()
-        }
-        LibraryApiContractId::JavaCollectionConstructor(kind) => [
-            "ArrayList",
-            "java.util.ArrayList",
-            "LinkedList",
-            "java.util.LinkedList",
-        ]
-        .into_iter()
-        .filter_map(|type_name| library_java_collection_constructor_contract(lang, type_name, 0))
-        .filter(|contract| contract.id == LibraryApiContractId::JavaCollectionConstructor(kind))
-        .map(|contract| contract.callee)
-        .collect(),
-        LibraryApiContractId::JavaMapFactory(kind) => ["of", "ofEntries"]
-            .into_iter()
-            .filter_map(|method| library_java_map_factory_contract(lang, "Map", method))
-            .filter(|contract| contract.id == LibraryApiContractId::JavaMapFactory(kind))
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::JavaMapEntryFactory => {
-            library_java_map_entry_contract(lang, "Map", "entry")
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::RubySetFactory => {
-            library_ruby_set_factory_contract(lang, "Set", "new", 1)
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::JsLikeSetConstructor => {
-            library_js_like_set_constructor_contract(lang, "Set")
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::JsLikeMapConstructor => {
-            library_js_like_map_constructor_contract(lang, "Map")
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::MapKeyViewWrapper => {
-            library_map_key_view_wrapper_contract(lang, "Array", "from", 1)
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::JsLikeStaticIndexMembership(kind) => ["indexOf", "findIndex"]
-            .into_iter()
-            .filter_map(|method| library_static_index_membership_contract(lang, method, 1))
-            .filter(|contract| {
-                contract.id == LibraryApiContractId::JsLikeStaticIndexMembership(kind)
-            })
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::MapGet => ["get"]
-            .into_iter()
-            .filter_map(|method| {
-                library_map_get_contract(lang, method, 1).map(|contract| contract.callee)
-            })
-            .collect(),
-        LibraryApiContractId::MapKeyView(kind) => ["keys", "keySet"]
-            .into_iter()
-            .filter_map(|method| library_map_key_view_contract(lang, method, 0))
-            .filter(|contract| contract.result.kind == kind)
-            .map(|contract| contract.callee)
-            .collect(),
-        LibraryApiContractId::IteratorIdentityAdapter => {
-            let methods = [
-                "iter",
-                "into_iter",
-                "iter_mut",
-                "collect",
-                "to_vec",
-                "copied",
-                "cloned",
-                "stream",
-            ];
-            methods
-                .into_iter()
-                .filter_map(|method| {
-                    library_iterator_identity_adapter_contract(lang, method, 0)
-                        .map(|contract| contract.callee)
-                })
-                .collect()
-        }
-        LibraryApiContractId::StaticCollectionAdapter => {
-            library_static_collection_adapter_contract(lang, "Arrays", "stream", 1)
-                .map(|contract| vec![contract.callee])
-                .unwrap_or_default()
-        }
-        LibraryApiContractId::MethodCall(semantic) => {
-            method_call_contract_callees_for_semantic(lang, semantic)
-        }
-        _ => Vec::new(),
-    }
-}
-
-fn library_free_function_builtin_callee_contracts_for_id(
-    lang: Lang,
-    builtin: Builtin,
-) -> Vec<LibraryApiCalleeContract> {
-    let candidate = match (lang, builtin) {
-        (Lang::Python, Builtin::Len) => Some(("len", 1)),
-        (Lang::Go, Builtin::Len) => Some(("len", 1)),
-        (Lang::Go, Builtin::Append) => Some(("append", 2)),
-        (Lang::Python, Builtin::Print) => Some(("print", 0)),
-        (Lang::Python, Builtin::Range) => Some(("range", 1)),
-        (Lang::Python, Builtin::Sum) => Some(("sum", 1)),
-        (Lang::Python, Builtin::Min) => Some(("min", 1)),
-        (Lang::Python, Builtin::Max) => Some(("max", 1)),
-        (Lang::Python, Builtin::Abs) => Some(("abs", 1)),
-        (Lang::Python, Builtin::Zip) => Some(("zip", 2)),
-        (Lang::Python, Builtin::Enumerate) => Some(("enumerate", 1)),
-        (Lang::Python, Builtin::Any) => Some(("any", 1)),
-        (Lang::Python, Builtin::All) => Some(("all", 1)),
-        _ => None,
-    };
-    candidate
-        .and_then(|(name, arg_count)| library_free_function_builtin_contract(lang, name, arg_count))
-        .map(|contract| vec![contract.callee])
-        .unwrap_or_default()
-}
-
-fn method_call_contract_callees_for_semantic(
-    lang: Lang,
-    semantic: MethodSemanticContract,
-) -> Vec<LibraryApiCalleeContract> {
-    let methods = [
-        "append",
-        "push",
-        "log",
-        "info",
-        "debug",
-        "Println",
-        "Printf",
-        "Print",
-        "Abs",
-        "HasPrefix",
-        "HasSuffix",
-        "Contains",
-        "len",
-        "size",
-        "length",
-        "is_empty",
-        "isEmpty",
-        "empty?",
-        "nil?",
-        "is_none",
-        "is_some",
-        "startsWith",
-        "startswith",
-        "starts_with",
-        "start_with?",
-        "endsWith",
-        "endswith",
-        "ends_with",
-        "end_with?",
-        "containsKey",
-        "contains_key",
-        "key?",
-        "has_key?",
-        "__contains__",
-        "includes",
-        "include?",
-        "member?",
-        "contains",
-        "has",
-        "join",
-        "get",
-        "fetch",
-        "getOrDefault",
-        "unwrap_or",
-        "unwrap_or_else",
-        "map_or",
-        "reduce",
-        "Min",
-        "Max",
-        "abs",
-        "min",
-        "max",
-        "zip",
-        "fold",
-        "inject",
-        "map",
-        "collect",
-        "filter",
-        "select",
-        "flatMap",
-        "flat_map",
-        "filter_map",
-        "some",
-        "every",
-        "all",
-        "any",
-        "all?",
-        "any?",
-        "allMatch",
-        "anyMatch",
-        "sum",
-        "count",
-    ];
-    methods
-        .into_iter()
-        .flat_map(|method| {
-            (0..=3).filter_map(move |arity| library_method_call_contract(lang, method, arity))
-        })
-        .filter(|contract| contract.result.semantic == semantic)
-        .map(|contract| contract.callee)
-        .collect()
-}
-
 fn dependency_ids_are_present(record: &EvidenceRecord, dependencies: &[EvidenceId]) -> bool {
     dependencies
         .iter()
@@ -2965,823 +2460,4 @@ fn dependency_has_asserted_record(
                 && dependency.kind == kind
         })
     })
-}
-
-pub fn library_free_name_collection_factory_contract(
-    lang: Lang,
-    name: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    FREE_NAME_COLLECTION_FACTORIES
-        .iter()
-        .find(|row| row.lang.is_none_or(|row_lang| row_lang == lang) && row.names.contains(&name))
-        .and_then(|row| {
-            let matched_name = row
-                .names
-                .iter()
-                .copied()
-                .find(|candidate| *candidate == name)?;
-            let id = match lang {
-                Lang::Python => LibraryApiContractId::PythonBuiltinCollectionFactory,
-                Lang::Rust => LibraryApiContractId::RustStdCollectionFactory,
-                _ => return None,
-            };
-            Some(LibraryCollectionFactoryContract {
-                id,
-                callee: LibraryApiCalleeContract::FreeName {
-                    name: matched_name,
-                    shadow: library_free_name_shadow_policy(lang, row.shadow_guard),
-                },
-                result: LibraryCollectionFactoryResult::SequenceArgument,
-            })
-        })
-}
-
-pub fn library_free_name_collection_factory_contracts(
-    lang: Lang,
-) -> impl Iterator<Item = LibraryCollectionFactoryContract> {
-    FREE_NAME_COLLECTION_FACTORIES
-        .iter()
-        .filter(move |row| row.lang.is_none_or(|row_lang| row_lang == lang))
-        .flat_map(move |row| {
-            row.names
-                .iter()
-                .filter_map(move |name| library_free_name_collection_factory_contract(lang, name))
-        })
-}
-
-pub fn library_free_function_builtin_contract(
-    lang: Lang,
-    name: &str,
-    arg_count: usize,
-) -> Option<LibraryFreeFunctionBuiltinContract> {
-    let result = free_function_builtin_contract(lang, name, arg_count)?;
-    Some(LibraryFreeFunctionBuiltinContract {
-        id: LibraryApiContractId::FreeFunctionBuiltin(result.builtin),
-        callee: LibraryApiCalleeContract::FreeName {
-            name: result.name,
-            shadow: library_free_name_shadow_policy(lang, result.requires_unshadowed),
-        },
-        result,
-    })
-}
-
-pub fn library_imported_collection_factory_contract(
-    lang: Lang,
-    module: &str,
-    exported: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    IMPORTED_COLLECTION_FACTORIES
-        .iter()
-        .find(|row| {
-            row.lang.is_none_or(|row_lang| row_lang == lang)
-                && row.module == module
-                && row.exported == exported
-        })
-        .map(|row| LibraryCollectionFactoryContract {
-            id: LibraryApiContractId::PythonImportedCollectionFactory,
-            callee: LibraryApiCalleeContract::ImportedBinding {
-                module: row.module,
-                exported: row.exported,
-            },
-            result: LibraryCollectionFactoryResult::SequenceArgument,
-        })
-}
-
-pub fn library_imported_collection_factory_contracts(
-    lang: Lang,
-) -> impl Iterator<Item = LibraryCollectionFactoryContract> {
-    IMPORTED_COLLECTION_FACTORIES
-        .iter()
-        .filter(move |row| row.lang.is_none_or(|row_lang| row_lang == lang))
-        .filter_map(move |row| {
-            library_imported_collection_factory_contract(lang, row.module, row.exported)
-        })
-}
-
-pub fn library_free_name_map_factory_contract(
-    lang: Lang,
-    name: &str,
-) -> Option<LibraryMapFactoryContract> {
-    FREE_NAME_MAP_FACTORIES
-        .iter()
-        .find(|row| row.lang.is_none_or(|row_lang| row_lang == lang) && row.names.contains(&name))
-        .and_then(|row| {
-            let matched_name = row
-                .names
-                .iter()
-                .copied()
-                .find(|candidate| *candidate == name)?;
-            let id = match lang {
-                Lang::Rust => LibraryApiContractId::RustStdMapFactory,
-                _ => return None,
-            };
-            Some(LibraryMapFactoryContract {
-                id,
-                callee: LibraryApiCalleeContract::FreeName {
-                    name: matched_name,
-                    shadow: library_free_name_shadow_policy(lang, false),
-                },
-                result: LibraryMapFactoryResult::EntrySequence {
-                    entry_seq_tag: row.entry_seq_tag,
-                },
-            })
-        })
-}
-
-pub fn library_free_name_map_factory_contracts(
-    lang: Lang,
-) -> impl Iterator<Item = LibraryMapFactoryContract> {
-    FREE_NAME_MAP_FACTORIES
-        .iter()
-        .filter(move |row| row.lang.is_none_or(|row_lang| row_lang == lang))
-        .flat_map(move |row| {
-            row.names
-                .iter()
-                .filter_map(move |name| library_free_name_map_factory_contract(lang, name))
-        })
-}
-
-pub fn library_java_collection_factory_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    let contract = java_collection_factory_contract(lang, receiver, method)?;
-    Some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::JavaCollectionFactory(contract.kind),
-        callee: LibraryApiCalleeContract::JavaUtilStaticMember {
-            receiver: contract.receiver,
-            method: contract.method,
-        },
-        result: LibraryCollectionFactoryResult::VariadicElements {
-            single_arg_spreads_array: contract.single_arg_spreads_array,
-        },
-    })
-}
-
-pub fn library_java_collection_factory_contract_by_hash(
-    lang: Lang,
-    receiver: &str,
-    method_hash: u64,
-) -> Option<LibraryCollectionFactoryContract> {
-    ["of", "asList"].into_iter().find_map(|method| {
-        (stable_symbol_hash(method) == method_hash)
-            .then(|| library_java_collection_factory_contract(lang, receiver, method))
-            .flatten()
-    })
-}
-
-pub fn library_java_collection_constructor_contract(
-    lang: Lang,
-    type_name: &str,
-    arg_count: usize,
-) -> Option<LibraryCollectionFactoryContract> {
-    let contract = java_collection_constructor_contract(lang, type_name, arg_count)?;
-    Some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::JavaCollectionConstructor(contract.kind),
-        callee: LibraryApiCalleeContract::JavaUtilConstructor {
-            simple_type: contract.simple_type,
-            qualified_type: contract.qualified_type,
-            module: contract.module,
-            requires_import_for_simple_type: contract.requires_import_for_simple_type,
-            requires_no_local_type_shadow: contract.requires_no_local_type_shadow,
-        },
-        result: LibraryCollectionFactoryResult::EmptySequence,
-    })
-}
-
-pub fn library_java_map_factory_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-) -> Option<LibraryMapFactoryContract> {
-    let contract = java_map_factory_contract(lang, receiver, method)?;
-    Some(LibraryMapFactoryContract {
-        id: LibraryApiContractId::JavaMapFactory(contract.kind),
-        callee: LibraryApiCalleeContract::JavaUtilStaticMember {
-            receiver: contract.receiver,
-            method: contract.method,
-        },
-        result: LibraryMapFactoryResult::JavaFactory {
-            kind: contract.kind,
-        },
-    })
-}
-
-pub fn library_java_map_factory_contract_by_hash(
-    lang: Lang,
-    receiver: &str,
-    method_hash: u64,
-) -> Option<LibraryMapFactoryContract> {
-    ["of", "ofEntries"].into_iter().find_map(|method| {
-        (stable_symbol_hash(method) == method_hash)
-            .then(|| library_java_map_factory_contract(lang, receiver, method))
-            .flatten()
-    })
-}
-
-pub fn library_java_map_entry_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-) -> Option<LibraryMapEntryFactoryContract> {
-    java_map_entry_contract(lang, receiver, method).then_some(LibraryMapEntryFactoryContract {
-        id: LibraryApiContractId::JavaMapEntryFactory,
-        callee: LibraryApiCalleeContract::JavaUtilStaticMember {
-            receiver: "Map",
-            method: "entry",
-        },
-    })
-}
-
-pub fn library_java_map_entry_contract_by_hash(
-    lang: Lang,
-    receiver: &str,
-    method_hash: u64,
-) -> Option<LibraryMapEntryFactoryContract> {
-    (method_hash == stable_symbol_hash("entry"))
-        .then(|| library_java_map_entry_contract(lang, receiver, "entry"))
-        .flatten()
-}
-
-pub fn library_ruby_set_factory_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryCollectionFactoryContract> {
-    let contract = ruby_set_factory_contract(lang, receiver, method, arg_count)?;
-    Some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::RubySetFactory,
-        callee: LibraryApiCalleeContract::RubyRequireStaticMember {
-            receiver: contract.receiver,
-            method: contract.method,
-            required_module: contract.required_module,
-            shadow_root: contract.shadow_root,
-        },
-        result: LibraryCollectionFactoryResult::SequenceArgument,
-    })
-}
-
-pub fn library_ruby_set_factory_contract_by_hash(
-    lang: Lang,
-    receiver: &str,
-    method_hash: u64,
-    arg_count: usize,
-) -> Option<LibraryCollectionFactoryContract> {
-    (method_hash == stable_symbol_hash("new"))
-        .then(|| library_ruby_set_factory_contract(lang, receiver, "new", arg_count))
-        .flatten()
-}
-
-pub fn library_js_like_set_constructor_contract(
-    lang: Lang,
-    receiver: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    let contract = js_like_set_constructor_contract(lang, receiver)?;
-    Some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::JsLikeSetConstructor,
-        callee: LibraryApiCalleeContract::JsGlobalConstructor {
-            receiver: contract.receiver,
-            requires_unshadowed_global: contract.requires_unshadowed_global,
-        },
-        result: LibraryCollectionFactoryResult::StaticNonFloatSequenceArgument,
-    })
-}
-
-pub fn library_js_like_map_constructor_contract(
-    lang: Lang,
-    receiver: &str,
-) -> Option<LibraryMapFactoryContract> {
-    let contract = js_like_map_constructor_contract(lang, receiver)?;
-    Some(LibraryMapFactoryContract {
-        id: LibraryApiContractId::JsLikeMapConstructor,
-        callee: LibraryApiCalleeContract::JsGlobalConstructor {
-            receiver: contract.receiver,
-            requires_unshadowed_global: contract.requires_unshadowed_global,
-        },
-        result: LibraryMapFactoryResult::EntrySequence {
-            entry_seq_tag: contract.entry_seq_tag?,
-        },
-    })
-}
-
-pub fn library_rust_vec_macro_factory_contract(
-    lang: Lang,
-    name: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    (lang == Lang::Rust && name == "vec").then_some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::RustVecMacroFactory,
-        callee: LibraryApiCalleeContract::RustMacro {
-            name: "vec",
-            shadow: LibraryApiShadowPolicy::SameName,
-        },
-        result: LibraryCollectionFactoryResult::VariadicElements {
-            single_arg_spreads_array: false,
-        },
-    })
-}
-
-pub fn library_rust_vec_new_factory_contract(
-    lang: Lang,
-    name: &str,
-) -> Option<LibraryCollectionFactoryContract> {
-    let contract = rust_vec_new_factory_contract(lang, name)?;
-    Some(LibraryCollectionFactoryContract {
-        id: LibraryApiContractId::RustVecNewFactory,
-        callee: LibraryApiCalleeContract::FreeName {
-            name: match name {
-                "Vec::new" => "Vec::new",
-                "std::vec::Vec::new" => "std::vec::Vec::new",
-                "alloc::vec::Vec::new" => "alloc::vec::Vec::new",
-                _ => return None,
-            },
-            shadow: LibraryApiShadowPolicy::ExplicitRoot(contract.shadow_root),
-        },
-        result: LibraryCollectionFactoryResult::EmptySequence,
-    })
-}
-
-pub fn library_map_key_view_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryMapKeyViewContract> {
-    if arg_count != 0 {
-        return None;
-    }
-    let result = match (lang, method) {
-        (Lang::Python | Lang::Ruby, "keys") => MapKeyViewContract {
-            method: "keys",
-            kind: MapKeyViewKind::Collection,
-        },
-        (Lang::Java, "keySet") => MapKeyViewContract {
-            method: "keySet",
-            kind: MapKeyViewKind::Collection,
-        },
-        (Lang::JavaScript | Lang::TypeScript | Lang::Vue | Lang::Svelte | Lang::Html, "keys") => {
-            MapKeyViewContract {
-                method: "keys",
-                kind: MapKeyViewKind::Iterator,
-            }
-        }
-        _ => return None,
-    };
-    Some(LibraryMapKeyViewContract {
-        id: LibraryApiContractId::MapKeyView(result.kind),
-        callee: LibraryApiCalleeContract::Method {
-            method: result.method,
-            receiver: MethodReceiverContract::ExactMap,
-        },
-        result,
-    })
-}
-
-pub fn library_map_key_view_contract_by_hash(
-    lang: Lang,
-    method_hash: u64,
-    arg_count: usize,
-) -> Option<LibraryMapKeyViewContract> {
-    ["keys", "keySet"].into_iter().find_map(|method| {
-        (stable_symbol_hash(method) == method_hash)
-            .then(|| library_map_key_view_contract(lang, method, arg_count))
-            .flatten()
-    })
-}
-
-pub fn library_map_key_view_wrapper_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryMapKeyViewWrapperContract> {
-    if !js_like_lang(lang) || receiver != "Array" || method != "from" || arg_count != 1 {
-        return None;
-    }
-    let result = MapKeyViewWrapperContract {
-        receiver: "Array",
-        method: "from",
-        qualified_path: "Array.from",
-    };
-    Some(LibraryMapKeyViewWrapperContract {
-        id: LibraryApiContractId::MapKeyViewWrapper,
-        callee: LibraryApiCalleeContract::StaticGlobalMethod {
-            receiver: result.receiver,
-            method: result.method,
-            qualified_path: result.qualified_path,
-            requires_unshadowed_receiver: true,
-        },
-        result,
-    })
-}
-
-pub fn library_map_key_view_wrapper_contract_by_hash(
-    lang: Lang,
-    receiver: &str,
-    method_hash: u64,
-    arg_count: usize,
-) -> Option<LibraryMapKeyViewWrapperContract> {
-    (method_hash == stable_symbol_hash("from"))
-        .then(|| library_map_key_view_wrapper_contract(lang, receiver, "from", arg_count))
-        .flatten()
-}
-
-pub fn library_map_get_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryMapGetContract> {
-    if !matches!(
-        lang,
-        Lang::Java
-            | Lang::Rust
-            | Lang::JavaScript
-            | Lang::TypeScript
-            | Lang::Vue
-            | Lang::Svelte
-            | Lang::Html
-    ) || method != "get"
-        || arg_count != 1
-    {
-        return None;
-    }
-    let result = MapGetContract {
-        method: "get",
-        receiver: MethodReceiverContract::ExactMap,
-    };
-    Some(LibraryMapGetContract {
-        id: LibraryApiContractId::MapGet,
-        callee: LibraryApiCalleeContract::Method {
-            method: result.method,
-            receiver: result.receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_map_get_contract_by_hash(
-    lang: Lang,
-    method_hash: u64,
-    arg_count: usize,
-) -> Option<LibraryMapGetContract> {
-    (method_hash == stable_symbol_hash("get"))
-        .then(|| library_map_get_contract(lang, "get", arg_count))
-        .flatten()
-}
-
-pub fn library_js_array_is_array_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryStaticGlobalMethodContract> {
-    if !js_like_lang(lang) || receiver != "Array" || method != "isArray" || arg_count != 1 {
-        return None;
-    }
-    let result = StaticGlobalMethodContract {
-        receiver: "Array",
-        method: "isArray",
-        qualified_path: "Array.isArray",
-        requires_unshadowed_receiver: true,
-    };
-    Some(LibraryStaticGlobalMethodContract {
-        id: LibraryApiContractId::JsArrayIsArray,
-        callee: LibraryApiCalleeContract::StaticGlobalMethod {
-            receiver: result.receiver,
-            method: result.method,
-            qualified_path: result.qualified_path,
-            requires_unshadowed_receiver: result.requires_unshadowed_receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_js_boolean_coercion_contract(
-    lang: Lang,
-    function: &str,
-    arg_count: usize,
-) -> Option<LibraryStaticGlobalFunctionContract> {
-    if !js_like_lang(lang) || function != "Boolean" || arg_count != 1 {
-        return None;
-    }
-    let result = StaticGlobalFunctionContract {
-        function: "Boolean",
-        requires_unshadowed_function: true,
-    };
-    Some(LibraryStaticGlobalFunctionContract {
-        id: LibraryApiContractId::JsBooleanCoercion,
-        callee: LibraryApiCalleeContract::StaticGlobalFunction {
-            function: result.function,
-            requires_unshadowed_function: result.requires_unshadowed_function,
-        },
-        result,
-    })
-}
-
-pub fn library_regex_test_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryRegexTestContract> {
-    if !js_like_lang(lang) || method != "test" || arg_count != 1 {
-        return None;
-    }
-    let result = RegexTestContract {
-        method: "test",
-        required_receiver_fact: SourceFactKind::Literal(SourceLiteralKind::Regex),
-    };
-    Some(LibraryRegexTestContract {
-        id: LibraryApiContractId::RegexTest,
-        callee: LibraryApiCalleeContract::RegexLiteralMethod {
-            method: result.method,
-            required_receiver_fact: result.required_receiver_fact,
-        },
-        result,
-    })
-}
-
-pub fn library_static_index_membership_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryStaticIndexMembershipContract> {
-    let result = static_index_membership_contract(lang, method, arg_count)?;
-    Some(LibraryStaticIndexMembershipContract {
-        id: LibraryApiContractId::JsLikeStaticIndexMembership(result.kind),
-        callee: LibraryApiCalleeContract::StaticIndexMembershipMethod {
-            method: result.method,
-            receiver: result.receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_imported_namespace_function_contract(
-    lang: Lang,
-    function: &str,
-    arg_count: usize,
-) -> Option<LibraryImportedNamespaceFunctionContract> {
-    let result = match (lang, function, arg_count) {
-        (Lang::Python, "prod", 1 | 2) => ImportedNamespaceFunctionContract {
-            module: "math",
-            function: "prod",
-            receiver: MethodReceiverContract::ImportedNamespace("math"),
-            semantic: ImportedNamespaceFunctionSemantic::ProductReduction {
-                op: Op::Mul,
-                identity: 1,
-            },
-        },
-        _ => return None,
-    };
-    Some(LibraryImportedNamespaceFunctionContract {
-        id: LibraryApiContractId::ImportedNamespaceFunction(result.semantic),
-        callee: LibraryApiCalleeContract::ImportedNamespaceFunction {
-            module: result.module,
-            function: result.function,
-        },
-        result,
-    })
-}
-
-pub fn library_promise_then_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryPromiseThenContract> {
-    if !js_like_lang(lang) || method != "then" || arg_count != 1 {
-        return None;
-    }
-    let result = PromiseThenContract {
-        receiver: AsyncReceiverContract::ExactPromiseLike,
-    };
-    Some(LibraryPromiseThenContract {
-        id: LibraryApiContractId::PromiseThen,
-        callee: LibraryApiCalleeContract::AsyncMethod {
-            method: "then",
-            receiver: result.receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_iterator_identity_adapter_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryIteratorIdentityAdapterContract> {
-    let method = if lang == Lang::Rust && arg_count == 0 {
-        match method {
-            "iter" => "iter",
-            "into_iter" => "into_iter",
-            "iter_mut" => "iter_mut",
-            "collect" => "collect",
-            "to_vec" => "to_vec",
-            "copied" => "copied",
-            "cloned" => "cloned",
-            _ => return None,
-        }
-    } else if lang == Lang::Java && method == "stream" && arg_count == 0 {
-        "stream"
-    } else {
-        return None;
-    };
-    let result = IteratorIdentityAdapterContract {
-        receiver: IteratorAdapterReceiverContract::ExactIterableValue,
-    };
-    Some(LibraryIteratorIdentityAdapterContract {
-        id: LibraryApiContractId::IteratorIdentityAdapter,
-        callee: LibraryApiCalleeContract::IteratorAdapterMethod {
-            method,
-            receiver: result.receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_static_collection_adapter_contract(
-    lang: Lang,
-    receiver: &str,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryStaticCollectionAdapterContract> {
-    if lang != Lang::Java || receiver != "Arrays" || method != "stream" || arg_count != 1 {
-        return None;
-    }
-    let result = StaticCollectionAdapterContract {
-        module: "java.util",
-        exported: "Arrays",
-    };
-    Some(LibraryStaticCollectionAdapterContract {
-        id: LibraryApiContractId::StaticCollectionAdapter,
-        callee: LibraryApiCalleeContract::JavaUtilStaticMember {
-            receiver: result.exported,
-            method: "stream",
-        },
-        result,
-    })
-}
-
-pub fn library_method_call_contract(
-    lang: Lang,
-    name: &str,
-    arg_count: usize,
-) -> Option<LibraryMethodCallContract> {
-    let result = method_call_contract_shape(lang, name, arg_count)?;
-    let method = library_method_selector_name(name)?;
-    Some(LibraryMethodCallContract {
-        id: LibraryApiContractId::MethodCall(result.semantic),
-        callee: LibraryApiCalleeContract::Method {
-            method,
-            receiver: result.receiver,
-        },
-        result,
-    })
-}
-
-pub fn library_receiver_method_api_contract(
-    lang: Lang,
-    method: &str,
-    arg_count: usize,
-) -> Option<LibraryReceiverMethodApiContract> {
-    library_map_get_contract(lang, method, arg_count)
-        .map(|contract| LibraryReceiverMethodApiContract {
-            id: contract.id,
-            callee: contract.callee,
-            rule: "library_api_map_get",
-        })
-        .or_else(|| {
-            library_map_key_view_contract(lang, method, arg_count).map(|contract| {
-                LibraryReceiverMethodApiContract {
-                    id: contract.id,
-                    callee: contract.callee,
-                    rule: "library_api_map_key_view",
-                }
-            })
-        })
-        .or_else(|| {
-            library_iterator_identity_adapter_contract(lang, method, arg_count).map(|contract| {
-                LibraryReceiverMethodApiContract {
-                    id: contract.id,
-                    callee: contract.callee,
-                    rule: "library_api_iterator_identity_adapter",
-                }
-            })
-        })
-        .or_else(|| {
-            library_scalar_integer_method_contract(lang, method, arg_count).map(|contract| {
-                LibraryReceiverMethodApiContract {
-                    id: contract.id,
-                    callee: contract.callee,
-                    rule: "library_api_scalar_integer_method",
-                }
-            })
-        })
-        .or_else(|| {
-            library_rust_option_and_then_contract(lang, method, arg_count).map(|contract| {
-                LibraryReceiverMethodApiContract {
-                    id: contract.id,
-                    callee: contract.callee,
-                    rule: "library_api_rust_option_and_then",
-                }
-            })
-        })
-        .or_else(|| {
-            library_method_call_contract(lang, method, arg_count).map(|contract| {
-                LibraryReceiverMethodApiContract {
-                    id: contract.id,
-                    callee: contract.callee,
-                    rule: "library_api_method_call",
-                }
-            })
-        })
-}
-
-pub(crate) fn library_method_selector_name(name: &str) -> Option<&'static str> {
-    Some(match name {
-        "__contains__" => "__contains__",
-        "Abs" => "Abs",
-        "Contains" => "Contains",
-        "HasPrefix" => "HasPrefix",
-        "HasSuffix" => "HasSuffix",
-        "Max" => "Max",
-        "Min" => "Min",
-        "Print" => "Print",
-        "Printf" => "Printf",
-        "Println" => "Println",
-        "abs" => "abs",
-        "all" => "all",
-        "all?" => "all?",
-        "allMatch" => "allMatch",
-        "any" => "any",
-        "any?" => "any?",
-        "anyMatch" => "anyMatch",
-        "and_then" => "and_then",
-        "append" => "append",
-        "clamp" => "clamp",
-        "collect" => "collect",
-        "contains" => "contains",
-        "containsKey" => "containsKey",
-        "contains_key" => "contains_key",
-        "count" => "count",
-        "debug" => "debug",
-        "empty?" => "empty?",
-        "end_with?" => "end_with?",
-        "endsWith" => "endsWith",
-        "ends_with" => "ends_with",
-        "endswith" => "endswith",
-        "every" => "every",
-        "fetch" => "fetch",
-        "filter" => "filter",
-        "filter_map" => "filter_map",
-        "flatMap" => "flatMap",
-        "flat_map" => "flat_map",
-        "fold" => "fold",
-        "get" => "get",
-        "getOrDefault" => "getOrDefault",
-        "has" => "has",
-        "has_key?" => "has_key?",
-        "include?" => "include?",
-        "includes" => "includes",
-        "info" => "info",
-        "inject" => "inject",
-        "isEmpty" => "isEmpty",
-        "is_empty" => "is_empty",
-        "is_none" => "is_none",
-        "is_some" => "is_some",
-        "join" => "join",
-        "key?" => "key?",
-        "len" => "len",
-        "length" => "length",
-        "log" => "log",
-        "map" => "map",
-        "map_or" => "map_or",
-        "max" => "max",
-        "member?" => "member?",
-        "min" => "min",
-        "nil?" => "nil?",
-        "push" => "push",
-        "reduce" => "reduce",
-        "select" => "select",
-        "size" => "size",
-        "some" => "some",
-        "start_with?" => "start_with?",
-        "startsWith" => "startsWith",
-        "starts_with" => "starts_with",
-        "startswith" => "startswith",
-        "sum" => "sum",
-        "unwrap_or" => "unwrap_or",
-        "unwrap_or_else" => "unwrap_or_else",
-        "zip" => "zip",
-        _ => return None,
-    })
-}
-
-fn library_free_name_shadow_policy(lang: Lang, shadow_guard: bool) -> LibraryApiShadowPolicy {
-    if shadow_guard {
-        LibraryApiShadowPolicy::SameName
-    } else if lang == Lang::Rust {
-        LibraryApiShadowPolicy::RustStdRootForStdPath
-    } else {
-        LibraryApiShadowPolicy::None
-    }
 }
